--- js/posts/index.js
@@ -1,3 +1,0 @@
-export * from "./feed.js";
-export * from "./createNewPost.js";
-export * from "./showPostById.js";
