+++ js/ui/index.js
@@ -1,0 +1,2 @@
+export * from "./posts/feed.js";
+export * from "./posts/createNewPost.js";
