+++ js/ui/posts/createNewPost.js
@@ -1,6 +1,6 @@
-import { BASE_URL } from "../api/constants.js";
-import { sendNewPost } from "../api/posts/sendNewPost.js";
-import { accessToken } from "../api/token.js";
+import { BASE_URL } from "../../api/constants.js";
+import { sendNewPost } from "../../api/posts/sendNewPost.js";
+import { accessToken } from "../../api/token.js";
 
 /**
  * Create new post and make an API call
